refactor: 💡 fix folder structure and add unified stories
AUTHORS: WILLIAM BRUCE AND MARCUS HOGLER
