--- src/Molecules/DatePicker/Calendar/constants.ts
@@ -1,4 +1,0 @@
-export const NUMBER_OF_VISIBLE_WEEKS_SINGLE = 6;
-export const NUMBER_OF_VISIBLE_WEEKDAYS_SINGLE = 7;
-export const NUMBER_OF_VISIBLE_ROWS_DOUBLE = 6;
-export const NUMBER_OF_VISIBLE_WEEKDAYS_DOUBLE = 14;
