--- src/Molecules/DatePicker/Calendar/index.ts
@@ -1,4 +1,0 @@
-import Calendar from './Calendar';
-import DoubleCalendar from './DoubleCalendar';
-
-export { Calendar, DoubleCalendar };
