+++ src/Molecules/DatePicker/Double/Calendar/index.ts
@@ -1,0 +1,1 @@
+export { default } from './DoubleCalendar';
