+++ src/Molecules/DatePicker/Double/DoubleDatePicker.types.ts
@@ -1,0 +1,22 @@
+export type Props = {
+  ariaLabelPrevious?: string;
+  ariaLabelNext?: string;
+  open?: boolean;
+  onChange?: (startDate: Date | null, endDate?: Date | null) => void;
+  label: string;
+  disableDate?: (date: Date) => boolean;
+  disabled?: boolean;
+  enableDate?: (date: Date) => boolean;
+  id: string;
+  selectedStartDate?: Date;
+  selectedEndDate?: Date;
+  inputValueStart?: string;
+  inputValueEnd?: string;
+  inputSize?: 's';
+  variant: 'DOUBLE';
+  width?: number;
+  ref?: React.Ref<HTMLDivElement>;
+  yearSelectLength?: number;
+  as?: keyof JSX.IntrinsicElements | React.ComponentType<any>;
+  disallowSingleDayRange?: boolean;
+};
