--- src/Molecules/DatePicker/Header/index.ts
@@ -1,4 +1,0 @@
-import Header from './Header';
-import DoubleHeader from './DoubleHeader';
-
-export { Header, DoubleHeader };
