+++ src/Molecules/DatePicker/Single/Calendar/Calendar.types.ts
@@ -1,20 +1,3 @@
-export type CalendarDayProps = {
-  className?: string;
-  date: Date;
-  disabled?: boolean;
-  enabled?: boolean;
-  locale: any;
-  onClick?: (date: Date) => void;
-  onKeyDown: (event: React.KeyboardEvent) => void;
-  onFocus: () => void;
-  sameMonth?: boolean;
-  selected?: boolean;
-  isWithinRange?: boolean;
-  isFirstDay?: boolean;
-  isLastDay?: boolean;
-  withGutter?: boolean;
-};
-
 type FocusedState = [number | null, number | null];
 
 export type Props = {
