+++ src/Molecules/DatePicker/Single/Header/Header.types.ts
@@ -1,0 +1,10 @@
+export type Props = {
+  ariaLabelPrevious?: string;
+  ariaLabelNext?: string;
+  id: string;
+  locale: string;
+  viewedDate: Date;
+  onMonthChange: (index: number) => void;
+  onYearChange: (year: number) => void;
+  yearSelectLength?: number;
+};
