+++ src/Molecules/DatePicker/index.ts
@@ -1,3 +1,3 @@
-import { DatePicker } from './DatePicker';
+import DatePicker from './DatePicker';
 
 export default DatePicker;
