+++ src/Molecules/DatePicker/shared/components/CalendarDay/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CalendarDay';
