+++ src/Molecules/DatePicker/shared/components/SelectMonth/SelectMonth.tsx
@@ -1,15 +1,15 @@
 import React, { useMemo, useState } from 'react';
 import format from 'date-fns/format';
 import styled, { useTheme } from 'styled-components';
-import { getLocale } from '../shared/dateUtils';
+import { getLocale } from '../../dateUtils';
 import { Props } from './SelectMonth.types';
 
 /**
  * Imported seperately because when imported in src/index.ts, Input will not have been imported yet and error will be thrown
  */
-import Input from '../../Input';
-import { Box, Icon, Flexbox, Typography } from '../../..';
-import { capitalize } from '../shared/textUtils';
+import Input from '../../../../Input';
+import { Box, Icon, Flexbox, Typography } from '../../../../..';
+import { capitalize } from '../../textUtils';
 
 const months = [...Array(12).keys()];
